Switch to testify/assert.
This keeps the invocation and style of the stock Go testing library but
reduces the verbosity of the assertions themselves. It can be used
interchangeably with the stock assertions. This seems easier to read to
me and requires very little learning beyond what you'd already have to
learn to write Go tests.