Unexport many things that are really internal.
Golint clued me in to how much I had exported that I had no intention of
writing documentation for because it was internal. It made me realize
that a lot of stuff with Capitol names that I just default to doing that
way (structs that are class-like, constants) didn't need to be exported.
It still looks weird to me, but it's the right thing to do, I think.

This spec was sort of combined with thinking about how I will write the
GoDoc (without actually doing it), and so there is a new stub for
package documentation, and a few TODOs that capture known issues are now
BUGs instead.