Remove any validations that aren't absolutely required.
We started with basic stuff, like making sure your server had an ID
before we tried to fetch it. Then we started creating servers and there
were a few more required fields. Then I realized that there were even
more, and that there were complicated validations like the name not
being more than 7 characters when combined with the account alias. What
if you put "seven" memoryGB instead of 7?

It's more than we should deal with. I am now handling the error response
that includes validation errors, and our docs should point straight
through to the Tier3 docs for the real skinny on what is required and
what is optional. We don't control and API and won't know when things
change anyway, even if we got it perfect now.

Some basic validations have to stay, like refusing to fetch a server
with no ID. If we don't even have enough info to form the URL, we're not
going to be able to return you any helpful error.