Make FetchCredentials use saveEntity.
This was built before that stuff existed, and can basically follow the
same pattern as everything else with a little tweaking.

I almost reverted this commit because it felt weird that one path
(statusProvidingEntity) has no JSON Unmarshalling and the other does,
but it sort of makes sense. If you return a status, you probably can't
return a representation of yourself yet. If you don't return a status,
you probably return your representation on create/update, since it's
ready to go right then. That fits with what we've done so far, anyway.