Use a bespoke test entity for GetEntity tests.
We were using Server, which implements the interface we need to test how
getEntity works. But we're relying on the real behavior of Server, like
how it will raise an error if the ID is empty. We're using this
side-effect of unrelated code to test the behavior of getEntity, which
is kind of lame. The testEntity struct implements the interface and
allows us to control entirely in the test how it should behave.