Move requestor and auth into client.
Requestor going into client was a no-brainer and let me get rid of
exported/unexported versions of Get/SaveEntity. I didn't anticipate
having to change the authentication stuff yet, but it was all kind of
tied up together and in a jenky state because of the client changes I
made (since the auth code calls SaveEntity on credentials but previously
existed outside of the context of a client). I'm happy with what I've
got now.