Move fixtures and example API server to a package.
I ran into an issue where fixtures used by the clcgo unit tests weren't
available to the tests in the examples. Either I needed to duplicate the
fixture data or move it into text files (and still have duplicated
helpers for loading those files), or I could move it into a package that
is useful primarily for testing purposes.

I didn't like the idea at first, but it's grown on me now that I'm
seeing the diff. That package is nice and self-contained, there aren't
fixture files hiding amongst the regular clcgo files, and the example
file is certainly shorter.