Fill out documentation.
While this doesn't satisfy lint yet, it does document all of the structs
and functions that I expect people to need to use. I'm struggling with
exporting or not exporting some of the interface functions, and those
aren't documented yet. Those are the bits that GoLint is so upset about.
I know I made them exported for a reason, but it's escaping me after the
long weekend.