Stop returning a *Status from save.
I was originally going to return a pointer because I was planning to do
a bunch of fancy logic where the status would be linked to whatever
entity it was reporting on and reload that entity. I never implemented
that, nor does it match any of the other return values. If I do
implement any fancier version of this library, it will probably exist as
a convenience package that uses the existing code to actually perform
the work.